flask_project/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.1
Flask-Limiter==3.12
python-dotenv==0.21.0
PyJWT==2.10.1
werkzeug==2.3.7
google-auth==2.40.3
together==1.5.17
cryptography==41.0.3
requests==2.31.0
python-magic==0.4.27
pypdf==3.17.0
python-docx==1.1.0
openpyxl==3.1.2
limits>=3.13.0,<4.0.0
ordered-set==4.1.0
typing-extensions==4.14.0
gunicorn==21.2.0
